services/django-api/requirements.txt
# =========================
# Django & REST
# =========================
Django==4.2.9
djangorestframework==3.14.0

# =========================
# Database
# =========================
psycopg2-binary==2.9.9

# =========================
# Celery & Redis
# =========================
celery==5.3.6
redis==5.0.1
django-celery-beat==2.5.0

# =========================
# Storage & S3/MinIO
# =========================
boto3==1.34.14
django-storages==1.14.2

# =========================
# Utilities
# =========================
python-dotenv==1.0.0
pydantic==2.5.3
structlog==24.1.0
gunicorn==21.2.0
whitenoise==6.11.0
django-cors-headers==4.3.1   # ✅ fixes frontend API calls

# =========================
# Google Drive API (API key only)
# =========================
google-api-python-client==2.70.0
